requirements.txt
catboost==1.2.3
gradio_client==0.12.0
google-generativeai==0.4.0
openai==1.14.2
numpy
pandas
matplotlib==3.8.3
nltk==3.8.1
keras==2.10.0
tensorflow==2.10.1
scikit-learn==1.4.1.post1
wordcloud==1.9.3
seaborn
plotly==5.19.0
python-dotenv==1.0.1
deep-translator==1.11.4
streamlit==1.32.0